Update validateTokenHandler.js
This way the code will first check for the presence of the token and respond with an error if it is missing or not properly formatted. If the token is present and valid, it will verify and decode, set req.user, and call next(). Otherwise, it will throw an error and send a 401 status response.